Change type size_t->int64_t

+++ src/query/v2/multiframe.cpp
@@ -26,7 +26,7 @@ namespace memgraph::query::v2 {
 // static_assert(std::forward_iterator<InvalidFramesPopulator::Iterator> &&
 //               std::equality_comparable<InvalidFramesPopulator::Iterator>);
 
-MultiFrame::MultiFrame(size_t size_of_frame, size_t number_of_frames, utils::MemoryResource *execution_memory)
+MultiFrame::MultiFrame(int64_t size_of_frame, size_t number_of_frames, utils::MemoryResource *execution_memory)
     : frames_(utils::pmr::vector<FrameWithValidity>(
           number_of_frames, FrameWithValidity(size_of_frame, execution_memory), execution_memory)) {
   MG_ASSERT(number_of_frames > 0);
